+++ js/utils.js
@@ -1,0 +1,15 @@
+const getRandomInt = function (min, max) {
+  if (min < 0 || min >= max) {
+    return 'Введено число меньше 0 либо задан неправильный интервал';
+  }
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+};
+
+getRandomInt(0, 100);
+
+const getStringLength = (string, maxLength) => string.length <= maxLength;
+getStringLength('find string length', 5);
+
+export {getRandomInt};
